requirements.txt
# Core ML/DL frameworks
torch>=2.1.0
flash-attn>=2.3.0
transformers>=4.30.0

# Data handling
anndata>=0.9.0
scanpy>=1.9.0
polars>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
h5py>=3.8.0

# Configuration and utilities
omegaconf>=2.3.0
pyyaml>=6.0
tqdm>=4.65.0
rich>=13.0.0

# Distributed training
accelerate>=0.21.0

# Logging and monitoring (optional)
wandb>=0.15.0
tensorboard>=2.13.0

# Evaluation metrics
scikit-learn>=1.3.0

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# GPU monitoring and optimization
nvidia-ml-py3>=7.352.0
pynvml>=11.4.1

# Development and testing
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Memory profiling (optional)
memory-profiler>=0.60.0
psutil>=5.9.0
